amélioration bash (suppression fichiers inutiles et refonte du crud améliorer simplicité finish

+++ fiefdouglou.sh
@@ -1,11 +1,18 @@
 rm -rf .vs
 rm -rf packages
 cd fiefdouglou
-rm -rf bin
-rm -rf obj
-cd Properties
-rm Settings.settings
-cd ../../
+	rm -rf bin
+	rm -rf obj
+	cd Properties
+		if test -f Settings.settings; then rm Settings.settings; fi
+	cd ../
+		if test -f .editorconfig; then rm .editorconfig; fi
+		if test -f packages.config; then rm packages.config; fi
+		if test -f fiefdouglouDataSet.cs; then rm fiefdouglouDataSet.cs; fi
+		if test -f fiefdouglouDataSet.xss; then rm fiefdouglouDataSet.xss; fi
+		if test -f fiefdouglouDataSet.xsd; then rm fiefdouglouDataSet.xsd; fi
+		if test -f fiefdouglouDataSet.xsc; then rm fiefdouglouDataSet.xsc; fi
+cd ../
 git status
 read -p "petite pause (y/n) : " -n 2 test
 echo $test
